requirements.txt
fastapi
uvicorn
llama-index
chromadb
python-dotenv
openai
pypdf